git: check-ref-format
